Affiche un contenu "aucun résultat" en fonction de la recherche et de l'onglet actif
